updated feature files
