+++ src/test/java/ApiTests/steps/pets/UpdatePet.java
@@ -1,4 +1,4 @@
-package ApiTests.Steps;
+package ApiTests.steps.pets;
 
 import cucumber.api.java.en.Given;
 import cucumber.api.java.en.Then;
@@ -18,21 +18,25 @@ import utils.models.pet.CreatePet;
 import utils.models.pet.PetData;
 import utils.models.pet.Tag;
 
-public class Pet {
+public class UpdatePet {
 
+    /**
+     *
+     */
+    
     private static Response res = null; // Response
     // Instantiate a Helper Test Methods (htm) Object
     HelperTestMethods htm = new HelperTestMethods();
     CreatePet createPet = new CreatePet();
 
 
-    @Given("^the user has POST api of pet$")
-public void the_user_has_POST_api_of_pet() throws Throwable {
+    @Given("^the user has PUT api of pet$")
+public void the_user_has_PUT_api_of_pet() throws Throwable {
     Utils.setBaseURI(); // Setup Base URI
 }
 
-@When("^the user enters the pet details$")
-public void the_user_enters_the_pet_details() throws Throwable {
+@When("^the user enters the new pet details$")
+public void the_user_enters_the_new_pet_details() throws Throwable {
     createPet.setId(1101);
 
         ObjectMapper mp = new ObjectMapper();
@@ -40,27 +44,27 @@ public void the_user_enters_the_pet_details() throws Throwable {
         Tag tag = mp.convertValue(PetData.tagData(), Tag.class);
 
         createPet.setCategory(ct);
-        createPet.setName("tackie");
+        createPet.setName("browny");
         ArrayList<String> photoUrls = new ArrayList<String>();
-        photoUrls.add("photoe");
-        photoUrls.add("photol");
+        photoUrls.add("photo1_updated");
+        photoUrls.add("photo2_upatd");
         createPet.setPhotoUrls(photoUrls);
 
         ArrayList<Tag> tags = new ArrayList<Tag>();
         tags.add(tag);
         createPet.setTags(tags);
-        createPet.setStatus("available");
+        createPet.setStatus("sold");
    
 }
 
-@When("^the user hits the api$")
-public void the_user_hits_the_api() throws Throwable {
-    res = Utils.getPetPostResponse(ContentType.JSON, createPet); // get post reponse
+@When("^the user hits the put api$")
+public void the_user_hits_the_put_api() throws Throwable {
+    res = Utils.getPetPutResponse(ContentType.JSON, createPet); // get put reponse
     Utils.getJsonPath(res); 
 }
 
-@Then("^the pet should be created$")
-public void the_pet_should_be_created() throws Throwable {
+@Then("^^the pet should be updated$")
+public void the_pet_should_be_updated() throws Throwable {
     htm.checkStatusIs200(res);
     String bodyStringValue = res.asString();
     Assert.assertTrue(bodyStringValue.contains("id"));
